API

MuJoCo.example_model_files — Method

Returns the list of absolute paths to available example model files provided by MuJoCo.

MuJoCo.example_model_files_directory — Method

Returns the absolute path of the root directory containing all the example model files.

MuJoCo.humanoid_model_file — Method

Returns the absolute path to the humanoid model provided by MuJoCo.

MuJoCo.init_data — Method

init_data(model::Model)

Creates an instance of the data required for the step! simulation.

Returns a Data object, wrapping the underlying mjData object.

MuJoCo.init_visualiser — Method

init_visualiser()

Loads the packages necessary for the running the visualiser.

Add the following packages to your project to be able to use the visualisation features, or run "install_visualiser()".

Packages:

FFMPEG

GLFW

Observables

PrettyTables

Printf

StaticArrays

MuJoCo.install_visualiser — Method

install_visualiser()

Installs the necessary packages for the running the visualiser into the current running environment.

Packages:

FFMPEG

GLFW

Observables

PrettyTables

Printf

StaticArrays

MuJoCo.load_model — Method

load_model(path)

Determines the type of file by the extension and loads the model into memory.

To use this model in a simulator, you will also need the corresponding data, obtained using init_data.

Expected files types: 'xml', 'mjcf', or 'mjb' (or uppercase variants).

Examples

model = load_model(MuJoCo.humanoid_model_file())
data = init_data(model)

MuJoCo.sample_model_and_data — Method

sample_model_and_data()

A utility module to create and initialise example Model and Data objects, reflecting the underlying mjModel and mjData structs to provide REPL code completition to aid development.

Returns a (model, data) tuple.

MuJoCo.step! — Method

step!(model::Model, data::Data)

Runs the simulation forward one time step, modifying the underlying data object.