MuJoCo.MJCore

MuJoCo.MJCore.mj_Euler — Method

mj_Euler(m::Model, d::Data)

Run Euler integrator, semi-implicit in velocity.

MuJoCo.MJCore.mj_RungeKutta — Method

mj_RungeKutta(m::Model, d::Data, N::Integer) -> Cvoid

Run Runge-Kutta explicit order-N integrator

MuJoCo.MJCore.mj_activate — Method

mj_activate(filename::String) -> Cint

Activate the license located at filename, return 1 if successful, otherwise throws a MuJoCoException.

MuJoCo.MJCore.mj_addContact — Method

mj_addContact(m::Model, d::Data, con::Ref{mjContact}) -> Cint

Add contact con to d.contact list. Return 0 if success; 1 if buffer full.

MuJoCo.MJCore.mj_addFileVFS — Method

mj_addFileVFS(vfs::Ref{mjVFS}, directory::String, filename::String) -> Cint

Add joinpath(directory, filename) to vfs, return 0: success, 1: full, 2: repeated name, -1: not found on disk.

MuJoCo.MJCore.mj_addM — Method

Add inertia matrix to destination matrix. Destination can be sparse uncompressed, or dense when all int* are NULL

MuJoCo.MJCore.mj_applyFT — Method

apply cartesian force and torque (outside xfrc_applied mechanism)

MuJoCo.MJCore.mj_camlight — Method

mj_camlight(m::Model, d::Data) -> Cvoid

Compute camera and light positions and orientations.

MuJoCo.MJCore.mj_certAnswer — Method

mj_certAnswer(question::MVector{16, mjtNum}, answer::MVector{16,mjtNum}) -> Cvoid

Generate client-side certificate answer given question and store it in answer.

MuJoCo.MJCore.mj_certCheck — Method

mj_certCheck(question::MVector{16,mjtNum}, answer::MVector{16,mjtNum}) -> Cint

Check certificate question-answer pair on server. Return 1 if match, 0 if mismatch.

MuJoCo.MJCore.mj_certQuestion — Method

mj_certQuestion(question::MVector{16,mjtNum}) -> Cvoid

Generate server-side certificate question and store it in question.

MuJoCo.MJCore.mj_checkAcc — Method

mj_checkAcc(m::Model, d::Data) -> Cvoid

Check qacc, reset if any element is too big or nan.

MuJoCo.MJCore.mj_checkPos — Method

mj_checkPos(m::Model, d::Data) -> Cvoid

Check qpos, reset if any element is too big or nan.

MuJoCo.MJCore.mj_checkVel — Method

mj_checkVel(m::Model, d::Data) -> Cvoid

Check qvel, reset if any element is too big or nan.

MuJoCo.MJCore.mj_collision — Method

mj_collision(m::Model, d::Data) -> Cvoid

Run collision detection.

MuJoCo.MJCore.mj_comPos — Method

mj_comPos(m::Model, d::Data) -> Cvoid

Map inertias and motion dofs to global frame centered at CoM.

MuJoCo.MJCore.mj_comVel — Method

mj_comVel(m::Model, d::Data) -> Cvoid

Compute cvel, cdof_dot.

MuJoCo.MJCore.mj_compareFwdInv — Method

mj_compareFwdInv(m::Model, d::Data) -> Cvoid

Compare forward and inverse dynamics, without changing results of forward dynamics. Saves results in d.fwdinv.

MuJoCo.MJCore.mj_constraintUpdate — Method

mj_constraintUpdate(m, d, jar, flg_coneHessian)
mj_constraintUpdate(m, d, jar, cost, flg_coneHessian)

Compute efc_state, efc_force, qfrc_constraint, and (optionally) cone Hessians. If specified, uses cost = s(jar) where jar = Jac*qacc-aref.

Arguments

m::Model

d::Data

jar::MJVec{mjtNum}

cost::MJVec{mjtNum}

flg_coneHessian::Integer

MuJoCo.MJCore.mj_contactForce — Method

extract 6D force:torque for one contact, in contact frame

MuJoCo.MJCore.mj_copyData — Method

mj_copyData(dest::Data, m::Model, src::Data) -> Ptr{mjData}

Copy mjData src into dest.

MuJoCo.MJCore.mj_copyModel — Method

mj_copyModel(dest::Model, src::Model) -> Ref{mjModel}

Copy model src into dest

MuJoCo.MJCore.mj_copyModel — Method

mj_copyModel(m::Model) -> Ptr{mjModel}

Create a copy of m, allocating new memory.

MuJoCo.MJCore.mj_crb — Method

mj_crb(m::Model, d::Data) -> Cvoid

Run composite rigid body inertia algorithm (CRB).

MuJoCo.MJCore.mj_deactivate — Method

mj_deactivate() -> Cvoid

Deactivate license and free all MuJoCo memory

MuJoCo.MJCore.mj_defaultLROpt — Method

mj_defaultLROpt(opt::Ref{mjLROpt}) -> Cvoid

Set default options for length range computation.

MuJoCo.MJCore.mj_defaultOption — Method

mj_defaultOption(opt::Ref{mjOption}) -> Cvoid

Set physics options to default values.

MuJoCo.MJCore.mj_defaultSolRefImp — Method

mj_defaultSolRefImp(solref::MJVec{mjtNum}, solimp::MJVec{mjtNum}) -> Cvoid

Set default solver parameters.

MuJoCo.MJCore.mj_defaultVFS — Method

mj_defaultVFS(vfs::Ref{mjVFS}) -> Cvoid

Initialize vfs to empty (no deallocation).

MuJoCo.MJCore.mj_defaultVisual — Method

mj_defaultVisual(vis::Ref{mjVisual}) -> Cvoid

Set visual options to default values.

MuJoCo.MJCore.mj_deleteData — Method

mj_deleteData(d::Ptr{mjData}) -> Nothiing

De-allocate d. Should only be called with d allocated by MuJoCo.

MuJoCo.MJCore.mj_deleteFileVFS — Method

mj_deleteFileVFS(vfs::Ref{mjVFS}, filename::String) -> Cint

Delete filename from vfs, return 0: success, -1: not found in VFS.

MuJoCo.MJCore.mj_deleteModel — Method

mj_deleteModel(m::Ptr{mjModel}) -> Cvoid

De-allocate model m. Should only be called with m allocated by MuJoCo.

MuJoCo.MJCore.mj_deleteVFS — Method

mj_deleteVFS(vfs::Ref{mjVFS}) -> Cvoid

Delete all files from vfs.

MuJoCo.MJCore.mj_differentiatePos — Method

compute velocity by finite-differencing two positions

MuJoCo.MJCore.mj_energyPos — Method

mj_energyPos(m::Model, d::Data) -> Cvoid

Evaluate position-dependent energy (potential).

MuJoCo.MJCore.mj_energyVel — Method

mj_energyVel(m::Model, d::Data) -> Cvoid

Evaluate velocity-dependent energy (kinetic).

MuJoCo.MJCore.mj_factorM — Method

mj_factorM(m::Model, d::Data) -> Cvoid

Compute sparse L'DL factorizaton of the inertia matrix

MuJoCo.MJCore.mj_findFileVFS — Method

mj_findFileVFS(vfs::Ref{mjVFS}, filename::String) -> Cint

Return file index of filename in vfs, or -1 if not found in vfs.

MuJoCo.MJCore.mj_forward — Method

mj_forward(m::Model, d::Data) -> Cvoid

Compute forward dynamics.

MuJoCo.MJCore.mj_forwardSkip — Method

mj_forwardSkip(m::Model, d::Data, skipstage::mjtStage, skipsensor::Bool) -> Cvoid

Compute forward dynamics, skipping parts of computation pipeline as specified by skipstage. Sensor computations can be skipped using skipsensor. See also: mjtStage

MuJoCo.MJCore.mj_freeLastXML — Method

mj_freeLastXML() -> Cvoid

Free last XML model if loaded. Called internally at each load.

MuJoCo.MJCore.mj_fullM — Method

convert sparse inertia matrix M into full matrix

MuJoCo.MJCore.mj_fwdAcceleration — Method

mj_fwdAcceleration(m::Model, d::Data) -> Cvoid

Add up all non-constraint forces, compute d.qacc_unc

MuJoCo.MJCore.mj_fwdActuation — Method

mj_fwdActuation(m::Model, d::Data) -> Cvoid

Compute actuator force d.qfrc_actuation.

MuJoCo.MJCore.mj_fwdConstraint — Method

mj_fwdConstraint(m::Model, d::Data) -> Cvoid

Run selected constraint solver

MuJoCo.MJCore.mj_fwdPosition — Method

mj_fwdPosition(m::Model, d::Data) -> Cvoid

Run position-dependent computations.

MuJoCo.MJCore.mj_fwdVelocity — Method

mj_fwdVelocity(m::Model, d::Data) -> Cvoid

Run velocity-dependent computations.

MuJoCo.MJCore.mj_getTotalmass — Method

sum all body masses

MuJoCo.MJCore.mj_id2name — Method

get name of object with specified id; 0: invalid type or id; type is mjtObj

MuJoCo.MJCore.mj_integratePos — Method

integrate position with given velocity

MuJoCo.MJCore.mj_invConstraint — Method

mj_invConstraint(m::Model, d::Data) -> Cvoid

Apply the analytical formula for inverse constraint dynamics.

MuJoCo.MJCore.mj_invPosition — Method

mj_invPosition(m::Model, d::Data) -> Cvoid

Run position-dependent computations in inverse dynamics.

MuJoCo.MJCore.mj_invVelocity — Method

mj_invVelocity(m::Model, d::Data) -> Cvoid

Run velocity-dependent computations in inverse dynamics.

MuJoCo.MJCore.mj_inverse — Method

mj_inverse(m::Model, d::Data) -> Cvoid

Compute inverse dynamics.

MuJoCo.MJCore.mj_inverseSkip — Method

mj_inverseSkip(m::Model, d::Data, skipstage::mjtStage, skipsensor::Bool) -> Cvoid

Compute inverse dynamics, skipping parts of computation pipeline as specified by skipstage. Sensor computations can be skipped using skipsensor. See also: mjtStage

MuJoCo.MJCore.mj_isDual — Method

mj_isDual(m::Model) -> Cint

Determine type of solver (PGS is dual, CG and Newton are primal).

MuJoCo.MJCore.mj_isPyramidal — Method

mj_isPyramidal(m::Model) -> Cint

Determine type of friction cone.

MuJoCo.MJCore.mj_isSparse — Method

mj_isSparse(m::Model) -> Cint

Determine type of constraint Jacobian.

MuJoCo.MJCore.mj_jac — Method

mj_jac(m, d, jacp, jacr, point, body) -> Cvoid

Compute 3/6-by-nv Jacobian of global point attached to body.

Arguments

m::Model

d::Data

jacp::MJVec{mjtNum}

jacr::MJVec{mjtNum}

point::SizedMJVec{3,mjtNum}:

body::Integer

MuJoCo.MJCore.mj_jacBody — Method

mj_jacBody(m, d, jacp, jacr, point, body) -> Cvoid

Compute body frame end-effector Jacobian.

Arguments

m::Model

d::Data

jacp::MJVec

jacr::MJVec

point::SVector{3,mjtNum}

body::Integer

MuJoCo.MJCore.mj_jacBodyCom — Method

mj_jacBody(m, d, jacp, jacr, point, body) -> Cvoid

Compute body frame end-effector Jacobian.

Arguments

m::Model

d::Data

jacp::MJVec

jacr::MJVec

body::Integer

MuJoCo.MJCore.mj_jacGeom — Method

mj_jacGeom(m, d, jacp, jacr, point, geom) -> Cvoid

Compute body frame end-effector Jacobian.

Arguments

m::Model

d::Data

jacp::MJVec

jacr::MJVec

geom::Integer

MuJoCo.MJCore.mj_jacPointAxis — Method

Compute translation Jacobian of point, and rotation Jacobian of axis

MuJoCo.MJCore.mj_jacSite — Method

mj_jacSite(m, d, jacp, jacr, point, geom) -> Cvoid

Compute body frame end-effector Jacobian.

Arguments

m::Model

d::Data

jacp::MJVec

jacr::MJVec

site::Integer

MuJoCo.MJCore.mj_kinematics — Method

mj_kinematics(m::Model, d::Data) -> Cvoid

Run forward kinematics.

MuJoCo.MJCore.mj_loadModel — Method

mj_loadModel(filename::String) -> Ptr{mjModel}

Load binary MJB file from filename in vfs.

MuJoCo.MJCore.mj_loadModel — Method

mj_loadModel(filename::String) -> Ptr{mjModel}

Load binary MJB file from filename located on local filesystem.

MuJoCo.MJCore.mj_loadXML — Method

mj_loadXML(filename::String[, vfs::Ref{mjVFS}) -> Ptr{mjModel}

Parse filename XML file in MJCF or URDF format, compile it, return low-level model. If vfs is specified, look up files in vfs before checking disk. Throws MuJoCoException if failure.

MuJoCo.MJCore.mj_local2Global — Method

map from body local to global Cartesian coordinates

MuJoCo.MJCore.mj_makeConstraint — Method

mj_makeConstraint(m::Model, d::Data) -> Cvoid

Construct constraints.

MuJoCo.MJCore.mj_makeData — Method

mj_makeData(m::Model) -> Ptr{mjData}

Allocate mjData correponding to m.

MuJoCo.MJCore.mj_makeEmptyFileVFS — Method

mj_makeEmptyFileVFS(vfs::Ref{mjVFS}, filename::String, filesize::Integer) -> Cint

Make empty file filename of filesize bytes in vfs, return 0: success, 1: full, 2: repeated name.

MuJoCo.MJCore.mj_mulJacTVec — Method

mj_mulJacTVec(m::Model, d::Data, res::MJVec{mjtNum}, vec::MJVec{mjtNum}) -> Cvoid

Multiply dense or sparse constraint Jacobian transpose by vector. See also: mj_mulJacVec

MuJoCo.MJCore.mj_mulJacVec — Method

mj_mulJacVec(m::Model, d::Data, res::MJVec{mjtNum}, vec::MJVec{mjtNum}) -> Cvoid

Multiply dense or sparse constraint Jacobian by vector. See also: mj_mulJacTVec

MuJoCo.MJCore.mj_mulM — Method

multiply vector by inertia matrix

MuJoCo.MJCore.mj_mulM2 — Method

Multiply vector by (inertia matrix)^(1/2).

MuJoCo.MJCore.mj_name2id — Method

get id of object with specified name; -1: not found; type is mjtObj

MuJoCo.MJCore.mj_normalizeQuat — Method

normalize all quaterions in qpos-type vector

MuJoCo.MJCore.mj_objectAcceleration — Method

compute object 6D acceleration in object-centered frame, world/local orientation

MuJoCo.MJCore.mj_objectVelocity — Method

compute object 6D velocity in object-centered frame, world/local orientation

MuJoCo.MJCore.mj_passive — Method

mj_passive(m::Model, d::Data) -> Cvoid

Compute qfrc_passive from spring-dampers, viscosity and density.

MuJoCo.MJCore.mj_printData — Method

mj_printData(m::Model, d::Data, filename::String) -> Cvoid

Print d to filename on local filesystem.

MuJoCo.MJCore.mj_printModel — Method

mj_printModel(m::Model, filename::String) -> Cvoid

Print m to filename on local filesystem.

MuJoCo.MJCore.mj_projectConstraint — Method

mj_projectConstraint(m::Model, d::Data) -> Cvoid

Compute inverse constaint inertia efc_AR.

MuJoCo.MJCore.mj_ray — Method

Intersect ray (pnt+x*vec, x>=0) with visible geoms, except geoms in bodyexclude. Return geomid and distance (x) to nearest surface, or -1 if no intersection. geomgroup, flg_static are as in mjvOption; geomgroup==NULL skips group exclusion.

MuJoCo.MJCore.mj_rayHfield — Method

Interect ray with hfield, return nearest distance or -1 if no intersection.

MuJoCo.MJCore.mj_rayMesh — Method

Interect ray with mesh, return nearest distance or -1 if no intersection.

MuJoCo.MJCore.mj_referenceConstraint — Method

mj_referenceConstraint(m::Model, d::Data) -> Cvoid

Compute efc_vel, efc_aref.

MuJoCo.MJCore.mj_resetCallbacks — Method

mj_resetCallbacks() -> Cvoid

Reset all callbacks to CNULL (CNULL is the default).

MuJoCo.MJCore.mj_resetData — Method

mj_resetData(m::Model, d::Data) -> Cvoid

Set data to defaults.

MuJoCo.MJCore.mj_resetDataDebug — Method

mj_resetDataDebug(m::Model, d::Data, debug_value::Cuchar) -> Cvoid

Set d to defaults, fill everything else with debug_value.

MuJoCo.MJCore.mj_resetDataKeyframe — Method

mj_resetDataKeyframe(m::Model, d::Data, key::Integer) -> Cvoid

Reset d and set fields from keyframe key.

MuJoCo.MJCore.mj_rne — Method

mj_rne(m::Model, d::Data, flg_acc::Integer, result::MJVec{mjtNum}) -> Cvoid

RNE: compute M(qpos)*qacc + C(qpos,qvel); flg_acc=0 removes inertial term.

MuJoCo.MJCore.mj_rnePostConstraint — Method

mj_rnePostConstraint(m::Model, d::Data) -> Cvoid

RNE with complete data: compute cacc, cfrcext, cfrcint.

MuJoCo.MJCore.mj_saveLastXML — Method

mj_saveLastXML(filename::String, m::Model) -> Cint

Update XML data structures with info from low-level model m, save as MJCF to filename. Returns 1 if success, else throws a MuJoCoException.

MuJoCo.MJCore.mj_saveModel — Method

mj_saveModel(m::Model, filename::String) -> Cvoid

Save m to filename

MuJoCo.MJCore.mj_saveModel — Method

mj_saveModel(m::Model, buffer::MJVec{UInt8}) -> Cvoid

Save m to buffer.

MuJoCo.MJCore.mj_sensorAcc — Method

mj_sensorAcc(m::Model, d::Data) -> Cvoid

Evaluate acceleration and force-dependent sensors.

MuJoCo.MJCore.mj_sensorPos — Method

mj_sensorPos(m::Model, d::Data) -> Cvoid

Evaluate position-dependent sensors.

MuJoCo.MJCore.mj_sensorVel — Method

mj_sensorVel(m::Model, d::Data) -> Cvoid

Evaluate velcoity-dependent sensors.

MuJoCo.MJCore.mj_setConst — Method

mj_setConst(m::Model, d::Data) -> Cvoid

Set constant fields of mjModel, corresponding to qpos0 configuration.

MuJoCo.MJCore.mj_setLengthRange — Method

mj_setLengthRange(m::Model, d::Data, index::Integer, opt::Ref{mjLROpt})

Set actuator_lengthrange for specified actuator; return 1 if ok, otherwise throws MuJoCoException.

MuJoCo.MJCore.mj_setTotalmass — Method

scale body masses and inertias to achieve specified total mass

MuJoCo.MJCore.mj_sizeModel — Method

mj_sizeModel(m::Model) -> Cint

Size of buffer in bytes needed to hold m.

MuJoCo.MJCore.mj_solveM — Method

mj_solveM(m, d, x, y, n) -> Cvoid

Solve linear system M * x = y using factorization: x = inv(L'DL)*y. See also: mj_solveM2

Arguments

m::Model

d::Data

x::MJVec{mjtNum}

y::MJVec{mjtNum}

n::Integer

MuJoCo.MJCore.mj_solveM2 — Method

mj_solveM2(m, d, x, y, n) -> Cvoid

Half of linear solve: x = sqrt(inv(D))inv(L')y Same arguments as mj_solveM.

MuJoCo.MJCore.mj_step — Method

mj_step(m::Model, d::Data) -> Cvoid

Advance simulation. use control callback if specified, otherwise use controls in d.ctrl, d.xfrc_applied, d.qfrc_applied. RK4 available. See also: mj_step1, mj_step2

MuJoCo.MJCore.mj_step1 — Method

mj_step1(m::Model, d::Data) -> Cvoid

Advance simulation in two steps: before external force/control is set by user. See also: mj_step2, mj_step

MuJoCo.MJCore.mj_step2 — Method

mj_step2(m::Model, d::Data) -> Cvoid

Advance simulation in two steps: after external force/control is set by user. See also: mj_step1, mj_step

MuJoCo.MJCore.mj_subtreeVel — Method

mj_subtreeVel(m::Model, d::Data) -> Cvoid

Compute subtree linear velocity and angular momentum.

MuJoCo.MJCore.mj_tendon — Method

mj_tendon(m::Model, d::Data) -> Cvoid

Compute tendon lengths, velocities and moment arms.

MuJoCo.MJCore.mj_transmission — Method

mj_transmission(m::Model, d::Data) -> Cvoid

Compute actuator transmission lengths and moments.

MuJoCo.MJCore.mj_version — Method

version number: 1.0.2 is encoded as 102 #TODO comment??

MuJoCo.MJCore.mj_warning — Method

high-level warning function: count warnings in Data, print only the first

MuJoCo.MJCore.mjr_addAux — Method

Add Aux buffer with given index to context; free previous Aux buffer.

MuJoCo.MJCore.mjr_blitAux — Method

Blit from Aux buffer to con->currentBuffer.

MuJoCo.MJCore.mjr_blitBuffer — Method

blit from src viewpoint in current framebuffer to dst viewport in other framebuffer if src, dst have different size and flgdepth==0, color is interpolated with GLLINEAR

MuJoCo.MJCore.mjr_changeFont — Method

Change font of existing context.

MuJoCo.MJCore.mjr_defaultContext — Method

set default mjrContext

MuJoCo.MJCore.mjr_drawPixels — Method

draw pixels from client buffer to current OpenGL framebuffer viewport is in OpenGL framebuffer; client buffer starts at (0,0)

MuJoCo.MJCore.mjr_figure — Method

draw lines

MuJoCo.MJCore.mjr_findRect — Method

Find first rectangle containing mouse, -1: not found.

MuJoCo.MJCore.mjr_finish — Method

call glFinish

MuJoCo.MJCore.mjr_freeContext — Method

free resources in custom OpenGL context, set to default

MuJoCo.MJCore.mjr_getError — Method

call glGetError and return result

MuJoCo.MJCore.mjr_makeContext — Method

allocate resources in custom OpenGL context; fontscale is mjtFontScale

MuJoCo.MJCore.mjr_maxViewport — Method

get maximum viewport for active buffer

MuJoCo.MJCore.mjr_overlay — Method

draw text overlay; font is mjtFont; gridpos is mjtGridPos

MuJoCo.MJCore.mjr_readPixels — Method

read pixels from current OpenGL framebuffer to client buffer viewport is in OpenGL framebuffer; client buffer starts at (0,0)

MuJoCo.MJCore.mjr_rectangle — Method

draw rectangle

MuJoCo.MJCore.mjr_render — Method

3D rendering

MuJoCo.MJCore.mjr_restoreBuffer — Method

Make con->currentBuffer current again.

MuJoCo.MJCore.mjr_setAux — Method

Set Aux buffer for custom OpenGL rendering (call restoreBuffer when done).

MuJoCo.MJCore.mjr_setBuffer — Method

set OpenGL framebuffer for rendering: mjFBWINDOW or mjFBOFFSCREEN if only one buffer is available, set that buffer and ignore framebuffer argument

MuJoCo.MJCore.mjr_text — Method

draw text at (x,y) in relative coordinates; font is mjtFont

MuJoCo.MJCore.mjr_uploadHField — Method

(re) upload height field to GPU

MuJoCo.MJCore.mjr_uploadMesh — Method

(re) upload mesh to GPU

MuJoCo.MJCore.mjr_uploadTexture — Method

(re) upload texture to GPU

MuJoCo.MJCore.mju_Halton — Method

Generate Halton sequence.

MuJoCo.MJCore.mju_axisAngle2Quat — Method

Convert axisAngle to quaternion.

MuJoCo.MJCore.mju_clearHandlers — Method

clear user error and memory handlers

MuJoCo.MJCore.mju_d2n — Method

convert from double to mjtNum

MuJoCo.MJCore.mju_decodePyramid — Method

convert pyramid representation to contact force

MuJoCo.MJCore.mju_derivQuat — Method

Compute time-derivative of quaternion, given 3D rotational velocity.

MuJoCo.MJCore.mju_encodePyramid — Method

convert contact force to pyramid representation

MuJoCo.MJCore.mju_error — Method

main error function; does not return to caller

MuJoCo.MJCore.mju_error_i — Method

error function mju_with int argument; msg is a printf format string

MuJoCo.MJCore.mju_error_s — Method

error function mju_with string argument

MuJoCo.MJCore.mju_f2n — Method

convert from float to mjtNum

MuJoCo.MJCore.mju_free — Method

free memory (with free() by default)

MuJoCo.MJCore.mju_insertionSort — Method

insertion sort, increasing order

MuJoCo.MJCore.mju_isBad — Method

return 1 if nan or abs(x)>mjMAXVAL, 0 otherwise

MuJoCo.MJCore.mju_isZero — Method

return 1 if all elements are 0

MuJoCo.MJCore.mju_malloc — Method

allocate memory; byte-align on 8; pad size to multiple of 8

MuJoCo.MJCore.mju_mat2Quat — Method

Convert 3D rotation matrix to quaterion.

MuJoCo.MJCore.mju_mulQuat — Method

Muiltiply quaternions.

MuJoCo.MJCore.mju_mulQuatAxis — Method

Muiltiply quaternion and axis.

MuJoCo.MJCore.mju_muscleBias — Method

Muscle passive force, prm = (range[2], force, scale, lmin, lmax, vmax, fpmax, fvmax).

MuJoCo.MJCore.mju_muscleDynamics — Method

Muscle activation dynamics, prm = (tauact, taudeact).

MuJoCo.MJCore.mju_muscleGain — Method

Muscle active force, prm = (range[2], force, scale, lmin, lmax, vmax, fpmax, fvmax).

MuJoCo.MJCore.mju_n2d — Method

convert from mjtNum to double

MuJoCo.MJCore.mju_n2f — Method

convert from mjtNum to float

MuJoCo.MJCore.mju_negQuat — Method

Conjugate quaternion, corresponding to opposite rotation.

MuJoCo.MJCore.mju_printMat — Method

print matrix to screen

MuJoCo.MJCore.mju_printMatSparse — Method

Print sparse matrix to screen.

MuJoCo.MJCore.mju_quat2Mat — Method

Convert quaternion to 3D rotation matrix.

MuJoCo.MJCore.mju_quat2Vel — Method

Convert quaternion (corresponding to orientation difference) to 3D velocity.

MuJoCo.MJCore.mju_quatIntegrate — Method

Integrate quaterion given 3D angular velocity.

MuJoCo.MJCore.mju_quatZ2Vec — Method

Construct quaternion performing rotation from z-axis to given vector.

MuJoCo.MJCore.mju_rayGeom — Method

Interect ray with pure geom, return nearest distance or -1 if no intersection.

MuJoCo.MJCore.mju_raySkin — Method

Interect ray with skin, return nearest vertex id.

MuJoCo.MJCore.mju_rotVecQuat — Method

Rotate vector by quaternion.

MuJoCo.MJCore.mju_springDamper — Method

integrate spring-damper analytically, return pos(dt)

MuJoCo.MJCore.mju_standardNormal — Method

standard normal random number generator (optional second number)

MuJoCo.MJCore.mju_str2Type — Method

convert type name to type id (mjtObj)

MuJoCo.MJCore.mju_strncpy — Method

Call strncpy, then set dst[n-1] = 0.

MuJoCo.MJCore.mju_subQuat — Method

Subtract quaternions, express as 3D velocity: qb*quat(res) = qa.

MuJoCo.MJCore.mju_type2Str — Method

convert type id (mjtObj) to type name

MuJoCo.MJCore.mju_warning — Method

main warning function; returns to caller

MuJoCo.MJCore.mju_warningText — Method

warning text

MuJoCo.MJCore.mju_warning_i — Method

warning function mju_with int argument

MuJoCo.MJCore.mju_warning_s — Method

warning function mju_with string argument

MuJoCo.MJCore.mju_writeLog — Method

Write [datetime, type: message] to MUJOCO_LOG.TXT.

MuJoCo.MJCore.mjv_addGeoms — Method

add geoms from selected categories

MuJoCo.MJCore.mjv_alignToCamera — Method

rotate 3D vec in horizontal plane by angle between (0,1) and (forwardx,forwardy)

MuJoCo.MJCore.mjv_applyPerturbForce — Method

set perturb force,torque in d->xfrc_applied, if selected body is dynamic

MuJoCo.MJCore.mjv_applyPerturbPose — Method

set perturb pos,quat in d->mocap when selected body is mocap, and in d->qpos otherwise d->qpos written only if flg_paused and subtree root for selected body has free joint

MuJoCo.MJCore.mjv_averageCamera — Method

Return the average of two OpenGL cameras.

MuJoCo.MJCore.mjv_cameraInModel — Method

get camera info in model space: average left and right OpenGL cameras

MuJoCo.MJCore.mjv_cameraInRoom — Method

get camera info in room space: average left and right OpenGL cameras

MuJoCo.MJCore.mjv_defaultCamera — Method

set default camera

MuJoCo.MJCore.mjv_defaultFigure — Method

Set default figure.

MuJoCo.MJCore.mjv_defaultOption — Method

set default visualization options

MuJoCo.MJCore.mjv_defaultPerturb — Method

set default perturbation

MuJoCo.MJCore.mjv_defaultScene — Method

Set default abstract scene.

MuJoCo.MJCore.mjv_freeScene — Method

free abstract scene

MuJoCo.MJCore.mjv_frustumHeight — Method

get frustum height at unit distance from camera; average left and right OpenGL cameras

MuJoCo.MJCore.mjv_initGeom — Method

Initialize given geom fields when not NULL, set the rest to their default values.

MuJoCo.MJCore.mjv_initPerturb — Method

copy perturb pos,quat from selected body; set scale for perturbation

MuJoCo.MJCore.mjv_makeConnector — Method

Set (type, size, pos, mat) for connector-type geom between given points. Assume that mjv_initGeom was already called to set all other properties.

MuJoCo.MJCore.mjv_makeLights — Method

Make list of lights.

MuJoCo.MJCore.mjv_makeScene — Method

allocate and init abstract scene

MuJoCo.MJCore.mjv_model2room — Method

transform pose from model to room space

MuJoCo.MJCore.mjv_moveCamera — Method

move camera with mouse; action is mjtMouse

MuJoCo.MJCore.mjv_moveModel — Method

move model with mouse; action is mjtMouse

MuJoCo.MJCore.mjv_movePerturb — Method

move perturb object with mouse; action is mjtMouse

MuJoCo.MJCore.mjv_room2model — Method

transform pose from room to model space

MuJoCo.MJCore.mjv_select — Method

Select geom or skin with mouse, return bodyid; -1: none selected.

MuJoCo.MJCore.mjv_updateCamera — Method

update camera only

MuJoCo.MJCore.mjv_updateScene — Method

update entire scene

MuJoCo.MJCore.mjv_updateSkin — Method

Update skins.