MuJoCo

MuJoCo.FieldInfo — Type

FieldInfo

Metadata for fields of the various MuJoCo structs. Sizes are declared in the same order as C/row-major order and should be reversed in Julia. Trailing/leading singleton dimensions are dropped.

Fields

mjsize: the MuJoCo C size declared

dsize: the dynamic size for pre-allocated, variable length fields (e.g. mjData.contact is declared as (nconmax, ) but should be viewed as (ncon,).

jlsize: The size for "jl" prefixed structs e.g. jlData. By default this is mjsize, but can be provide new sizes (e.g. mjData.xmat is (9, nbody)whilejlData.xmatis(:xmat, nbody)which allows for named indexing likejlData.xmat[:xz, :foobody]`