MuJoCo API

MuJoCo

MuJoCo.example_model_files

MuJoCo.example_model_files_directory

MuJoCo.forward!

MuJoCo.get_physics_state

MuJoCo.humanoid_model_file

MuJoCo.init_data

MuJoCo.init_visualiser

MuJoCo.install_visualiser

MuJoCo.load_model

MuJoCo.mj_addM

MuJoCo.mj_applyFT

MuJoCo.mj_array

MuJoCo.mj_constraintUpdate

MuJoCo.mj_differentiatePos

MuJoCo.mj_fullM

MuJoCo.mj_getState

MuJoCo.mj_integratePos

MuJoCo.mj_jac

MuJoCo.mj_jacBody

MuJoCo.mj_jacBodyCom

MuJoCo.mj_jacGeom

MuJoCo.mj_jacPointAxis

MuJoCo.mj_jacSite

MuJoCo.mj_jacSubtreeCom

MuJoCo.mj_mulJacTVec

MuJoCo.mj_mulJacVec

MuJoCo.mj_mulM

MuJoCo.mj_mulM2

MuJoCo.mj_normalizeQuat

MuJoCo.mj_ray

MuJoCo.mj_rne

MuJoCo.mj_setState

MuJoCo.mj_solveM

MuJoCo.mj_solveM2

MuJoCo.mj_zeros

MuJoCo.mjd_inverseFD

MuJoCo.mjd_transitionFD

MuJoCo.mju_L1

MuJoCo.mju_add

MuJoCo.mju_addScl

MuJoCo.mju_addTo

MuJoCo.mju_addToScl

MuJoCo.mju_band2Dense

MuJoCo.mju_bandMulMatVec

MuJoCo.mju_boxQP

MuJoCo.mju_cholFactor

MuJoCo.mju_cholFactorBand

MuJoCo.mju_cholSolve

MuJoCo.mju_cholSolveBand

MuJoCo.mju_cholUpdate

MuJoCo.mju_copy

MuJoCo.mju_d2n

MuJoCo.mju_decodePyramid

MuJoCo.mju_dense2Band

MuJoCo.mju_dot

MuJoCo.mju_encodePyramid

MuJoCo.mju_eye

MuJoCo.mju_f2n

MuJoCo.mju_fill

MuJoCo.mju_insertionSort

MuJoCo.mju_insertionSortInt

MuJoCo.mju_isZero

MuJoCo.mju_mulMatMat

MuJoCo.mju_mulMatMatT

MuJoCo.mju_mulMatTMat

MuJoCo.mju_mulMatTVec

MuJoCo.mju_mulMatVec

MuJoCo.mju_mulVecMatVec

MuJoCo.mju_n2d

MuJoCo.mju_n2f

MuJoCo.mju_norm

MuJoCo.mju_normalize

MuJoCo.mju_printMat

MuJoCo.mju_scl

MuJoCo.mju_sqrMatTD

MuJoCo.mju_sub

MuJoCo.mju_subFrom

MuJoCo.mju_sum

MuJoCo.mju_symmetrize

MuJoCo.mju_transpose

MuJoCo.mju_zero

MuJoCo.reset!

MuJoCo.resetkey!

MuJoCo.resetkey!

MuJoCo.sample_model_and_data

MuJoCo.set_physics_state!

MuJoCo.step!

MuJoCo.timestep

MuJoCo.example_model_files — Method

Returns the list of absolute paths to available example model files provided by MuJoCo.

MuJoCo.example_model_files_directory — Method

Returns the absolute path of the root directory containing all the example model files.

MuJoCo.forward! — Method

forward!(model::MODEL_TYPES, data::DATA_TYPES)

Same as step! but without integrating in time.

MuJoCo.get_physics_state — Method

get_physics_state(m::Model, d::Data)

Extract the state vector of a MuJoCo model.

The state vector is [joint positions, joint velocities, actuator states] with dimension (nq, nv, na).

See also mj_getState and set_physics_state!.

MuJoCo.humanoid_model_file — Method

Returns the absolute path to the humanoid model provided by MuJoCo.

MuJoCo.init_data — Method

init_data(model::Model)

Creates an instance of the data required for the step! simulation.

Returns a Data object, wrapping the underlying mjData object.

MuJoCo.init_visualiser — Method

init_visualiser()

Loads the packages necessary for the running the visualiser.

Add the following packages to your project to be able to use the visualisation features, or run "install_visualiser()".

Packages:

BangBang

FFMPEG

GLFW

Observables

PrettyTables

Printf

StaticArrays

MuJoCo.install_visualiser — Method

install_visualiser()

Installs the necessary packages for the running the visualiser into the current running environment.

Packages:

BangBang

FFMPEG

GLFW

Observables

PrettyTables

Printf

StaticArrays

MuJoCo.load_model — Method

load_model(path)

Determines the type of file by the extension and loads the model into memory.

To use this model in a simulator, you will also need the corresponding data, obtained using init_data.

Expected files types: 'xml', 'mjcf', or 'mjb' (or uppercase variants).

Examples

model = load_model(MuJoCo.humanoid_model_file())
data = init_data(model)

MuJoCo.mj_addM — Method

mj_addM(m, d, dst, rownnz, rowadr, colind)

Add inertia matrix to destination matrix. Destination can be sparse uncompressed, or dense when all int* are NULL

Arguments

m::Model -> Constant.

d::Data

dst::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

rownnz::Vector{Int32} -> A vector of variable size. Check additional info for sizes.

rowadr::Vector{Int32} -> A vector of variable size. Check additional info for sizes.

colind::Vector{Int32} -> A vector of variable size. Check additional info for sizes.

Additional Info

dst should be a vector, not a matrix.

rownnz should be a vector, not a matrix.

rowadr should be a vector, not a matrix.

colind should be a vector, not a matrix.

dst should be of size nM

rownnz should be of size nv

rowadr should be of size nv

colind should be of size nM

MuJoCo.mj_applyFT — Method

mj_applyFT(m, d, force, torque, point, body, qfrc_target)

Apply Cartesian force and torque (outside xfrc_applied mechanism).

Arguments

m::Model -> Constant.

d::Data

force::Vector{Float64} -> A vector of size 3. Constant.

torque::Vector{Float64} -> A vector of size 3. Constant.

point::Vector{Float64} -> A vector of size 3. Constant.

body::Int32

qfrc_target::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

Additional Info

force should be a vector of size 3

force should be a vector of size 3.

torque should be a vector of size 3

torque should be a vector of size 3.

point should be a vector of size 3

point should be a vector of size 3.

qfrc_target should be a vector, not a matrix.

qfrc_target should be of size nv

MuJoCo.mj_array — Method

mj_array([element_type=mjtNum], dims...)

Allocates an array compatible with the underlying MuJoCo C API.

The C API treats arrays as row-major, while by default arrays in Julia are column-major. This function will create an array which is accessed in a row-major way, but can be treated by a normal array in your Julia code.

Arguments

element_type: Defaults to mjtNum (typically Float64).

dims: The dimensionality of output array. Can either be a tuple of integers or a series of integers.

MuJoCo.mj_constraintUpdate — Method

mj_constraintUpdate(m, d, jar, cost, flg_coneHessian)

Compute efcstate, efcforce, qfrc_constraint, and (optionally) cone Hessians. If cost is not NULL, set cost = s(jar) where jar = Jacqacc-aref.

Arguments

m::Model -> Constant.

d::Data

jar::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

cost::Vector{Float64} -> An optional vector of size 1.

flg_coneHessian::Int32

Additional Info

jar should be a vector, not a matrix.

cost should be a vector of size 1

cost should be a vector of size 1.

size of jar should equal nefc

MuJoCo.mj_differentiatePos — Method

mj_differentiatePos(m, qvel, dt, qpos1, qpos2)

Compute velocity by finite-differencing two positions.

Arguments

m::Model -> Constant.

qvel::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

dt::Float64

qpos1::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

qpos2::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

qvel should be a vector, not a matrix.

qpos1 should be a vector, not a matrix.

qpos2 should be a vector, not a matrix.

qvel should be of size nv

qpos1 should be of size nq

qpos2 should be of size nq

MuJoCo.mj_fullM — Method

mj_fullM(m, dst, M)

Convert sparse inertia matrix M into full (i.e. dense) matrix.

Arguments

m::Model -> Constant.

dst::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

M::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

M should be a vector, not a matrix.

M should be of size nM

dst should be of shape (nv, nv)

MuJoCo.mj_getState — Method

mj_getState(m, d, state, spec)

Get state.

Arguments

m::Model -> Constant.

d::Data -> Constant.

state::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

spec::Int32

Additional Info

state should be a vector, not a matrix.

state size should equal mj_stateSize(m, spec)

MuJoCo.mj_integratePos — Method

mj_integratePos(m, qpos, qvel, dt)

Integrate position with given velocity.

Arguments

m::Model -> Constant.

qpos::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

qvel::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

dt::Float64

Additional Info

qpos should be a vector, not a matrix.

qvel should be a vector, not a matrix.

qpos should be of size nq

qvel should be of size nv

MuJoCo.mj_jac — Method

mj_jac(m, d, jacp, jacr, point, body)

Compute 3/6-by-nv end-effector Jacobian of global point attached to given body.

Arguments

m::Model -> Constant.

d::Data

jacp::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

jacr::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

point::Vector{Float64} -> A vector of size 3. Constant.

body::Int32

Additional Info

point should be a vector of size 3

point should be a vector of size 3.

jacp should be of shape (3, nv)

jacr should be of shape (3, nv)

MuJoCo.mj_jacBody — Method

mj_jacBody(m, d, jacp, jacr, body)

Compute body frame end-effector Jacobian.

Arguments

m::Model -> Constant.

d::Data

jacp::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

jacr::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

body::Int32

Additional Info

jacp should be of shape (3, nv)

jacr should be of shape (3, nv)

MuJoCo.mj_jacBodyCom — Method

mj_jacBodyCom(m, d, jacp, jacr, body)

Compute body center-of-mass end-effector Jacobian.

Arguments

m::Model -> Constant.

d::Data

jacp::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

jacr::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

body::Int32

Additional Info

jacp should be of shape (3, nv)

jacr should be of shape (3, nv)

MuJoCo.mj_jacGeom — Method

mj_jacGeom(m, d, jacp, jacr, geom)

Compute geom end-effector Jacobian.

Arguments

m::Model -> Constant.

d::Data

jacp::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

jacr::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

geom::Int32

Additional Info

jacp should be of shape (3, nv)

jacr should be of shape (3, nv)

MuJoCo.mj_jacPointAxis — Method

mj_jacPointAxis(m, d, jacp, jacr, point, axis, body)

Compute translation end-effector Jacobian of point, and rotation Jacobian of axis.

Arguments

m::Model -> Constant.

d::Data

jacp::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

jacr::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

point::Vector{Float64} -> A vector of size 3. Constant.

axis::Vector{Float64} -> A vector of size 3. Constant.

body::Int32

Additional Info

point should be a vector of size 3

point should be a vector of size 3.

axis should be a vector of size 3

axis should be a vector of size 3.

jacp should be of shape (3, nv)

jacr should be of shape (3, nv)

MuJoCo.mj_jacSite — Method

mj_jacSite(m, d, jacp, jacr, site)

Compute site end-effector Jacobian.

Arguments

m::Model -> Constant.

d::Data

jacp::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

jacr::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

site::Int32

Additional Info

jacp should be of shape (3, nv)

jacr should be of shape (3, nv)

MuJoCo.mj_jacSubtreeCom — Method

mj_jacSubtreeCom(m, d, jacp, body)

Compute subtree center-of-mass end-effector Jacobian.

Arguments

m::Model -> Constant.

d::Data

jacp::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

body::Int32

Additional Info

jacp should be of shape (3, nv)

MuJoCo.mj_mulJacTVec — Method

mj_mulJacTVec(m, d, res, vec)

Multiply dense or sparse constraint Jacobian transpose by vector.

Arguments

m::Model -> Constant.

d::Data

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

vec::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

res should be a vector, not a matrix.

vec should be a vector, not a matrix.

res should be of length nv

vec should be of length nefc

MuJoCo.mj_mulJacVec — Method

mj_mulJacVec(m, d, res, vec)

Multiply dense or sparse constraint Jacobian by vector.

Arguments

m::Model -> Constant.

d::Data

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

vec::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

res should be a vector, not a matrix.

vec should be a vector, not a matrix.

res should be of length nefc

vec should be of length nv

MuJoCo.mj_mulM — Method

mj_mulM(m, d, res, vec)

Multiply vector by inertia matrix.

Arguments

m::Model -> Constant.

d::Data -> Constant.

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

vec::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

res should be a vector, not a matrix.

vec should be a vector, not a matrix.

res should be of size nv

vec should be of size nv

MuJoCo.mj_mulM2 — Method

mj_mulM2(m, d, res, vec)

Multiply vector by (inertia matrix)^(1/2).

Arguments

m::Model -> Constant.

d::Data -> Constant.

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

vec::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

res should be a vector, not a matrix.

vec should be a vector, not a matrix.

res should be of size nv

vec should be of size nv

MuJoCo.mj_normalizeQuat — Method

mj_normalizeQuat(m, qpos)

Normalize all quaternions in qpos-type vector.

Arguments

m::Model -> Constant.

qpos::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

Additional Info

qpos should be a vector, not a matrix.

qpos should be of size nq

MuJoCo.mj_ray — Method

mj_ray(m, d, pnt, vec, geomgroup, flg_static, bodyexclude, geomid)

Intersect ray (pnt+x*vec, x>=0) with visible geoms, except geoms in bodyexclude. Return distance (x) to nearest surface, or -1 if no intersection and output geomid. geomgroup, flg_static are as in mjvOption; geomgroup==NULL skips group exclusion.

Arguments

m::Model -> Constant.

d::Data -> Constant.

pnt::Vector{Float64} -> A vector of size 3. Constant.

vec::Vector{Float64} -> A vector of size 3. Constant.

geomgroup::Vector{UInt8} -> An optional vector of size 6. Constant.

flg_static::UInt8

bodyexclude::Int32

geomid::Vector{Int32} -> A vector of size 1.

Additional Info

pnt should be a vector of size 3

pnt should be a vector of size 3.

vec should be a vector of size 3

vec should be a vector of size 3.

geomgroup should be a vector of size 6

geomgroup should be a vector of size 6.

geomid should be a vector of size 1

geomid should be a vector of size 1.

MuJoCo.mj_rne — Method

mj_rne(m, d, flg_acc, result)

RNE: compute M(qpos)*qacc + C(qpos,qvel); flg_acc=0 removes inertial term.

Arguments

m::Model -> Constant.

d::Data

flg_acc::Int32

result::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

Additional Info

result should be a vector, not a matrix.

result should have length nv

MuJoCo.mj_setState — Method

mj_setState(m, d, state, spec)

Set state.

Arguments

m::Model -> Constant.

d::Data

state::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

spec::Int32

Additional Info

state should be a vector, not a matrix.

state size should equal mj_stateSize(m, spec)

MuJoCo.mj_solveM — Method

mj_solveM(m, d, x, y)

Solve linear system M * x = y using factorization: x = inv(L'DL)*y

Arguments

m::Model -> Constant.

d::Data

x::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

y::Matrix{Float64} -> A matrix variable size. Check additional info for sizes. Constant.

MuJoCo.mj_solveM2 — Method

mj_solveM2(m, d, x, y)

Half of linear solve: x = sqrt(inv(D))inv(L')y

Arguments

m::Model -> Constant.

d::Data

x::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

y::Matrix{Float64} -> A matrix variable size. Check additional info for sizes. Constant.

MuJoCo.mj_zeros — Method

mj_zeros([element_type=mjtNum], dims...)

Allocates an array full of zeros, compatible with the underlying MuJoCo C API.

The C API treats arrays as row-major, while by default arrays in Julia are column-major. This function will create an array which is accessed in a row-major way, but can be treated by a normal array in your Julia code.

Arguments

element_type: The element type of the array. Defaults to mjtNum (typically Float64).

dims: The dimensionality of output array. Can either be a tuple of integers or a series of integers.

MuJoCo.mjd_inverseFD — Method

mjd_inverseFD(m, d, eps, flg_actuation, DfDq, DfDv, DfDa, DsDq, DsDv, DsDa, DmDq)

Finite differenced Jacobians of (force, sensors) = mjinverse(state, acceleration) All outputs are optional. Output dimensions (transposed w.r.t Control Theory convention): DfDq: (nv x nv) DfDv: (nv x nv) DfDa: (nv x nv) DsDq: (nv x nsensordata) DsDv: (nv x nsensordata) DsDa: (nv x nsensordata) DmDq: (nv x nM) single-letter shortcuts: inputs: q=qpos, v=qvel, a=qacc outputs: f=qfrcinverse, s=sensordata, m=qM notes: optionally computes mass matrix Jacobian DmDq flgactuation specifies whether to subtract qfrcactuator from qfrc_inverse

Arguments

m::Model -> Constant.

d::Data

eps::Float64

flg_actuation::UInt8

DfDq::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

DfDv::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

DfDa::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

DsDq::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

DsDv::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

DsDa::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

DmDq::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

Additional Info

DfDq should be of shape (nv, nv)

DfDv should be of shape (nv, nv)

DfDa should be of shape (nv, nv)

DsDq should be of shape (nv, nsensordata)

DsDv should be of shape (nv, nsensordata)

DsDa should be of shape (nv, nsensordata)

DmDq should be of shape (nv, nM)

MuJoCo.mjd_transitionFD — Method

mjd_transitionFD(m, d, eps, flg_centered, A, B, C, D)

Finite differenced transition matrices (control theory notation) d(x_next) = Adx + Bdu d(sensor) = Cdx + Ddu required output matrix dimensions: A: (2nv+na x 2nv+na) B: (2nv+na x nu) D: (nsensordata x 2nv+na) C: (nsensordata x nu)

Arguments

m::Model -> Constant.

d::Data

eps::Float64

flg_centered::UInt8

A::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

B::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

C::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

D::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

Additional Info

A should be of shape (2nv+na, 2nv+na)

B should be of shape (2*nv+na, nu)

C should be of shape (nsensordata, 2*nv+na)

D should be of shape (nsensordata, nu)

MuJoCo.mju_L1 — Method

mju_L1(vec)

Return L1 norm: sum(abs(vec)).

Arguments

vec::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

Additional Info

vec should be a vector, not a matrix.

MuJoCo.mju_add — Method

mju_add(res, vec1, vec2)

Set res = vec1 + vec2.

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

vec1::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

vec2::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

res should be a vector, not a matrix.

vec1 should be a vector, not a matrix.

vec2 should be a vector, not a matrix.

res and vec1 should have the same size

res and vec2 should have the same size

MuJoCo.mju_addScl — Method

mju_addScl(res, vec1, vec2, scl)

Set res = vec1 + vec2*scl.

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

vec1::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

vec2::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

scl::Float64

Additional Info

res should be a vector, not a matrix.

vec1 should be a vector, not a matrix.

vec2 should be a vector, not a matrix.

res and vec1 should have the same size

res and vec2 should have the same size

MuJoCo.mju_addTo — Method

mju_addTo(res, vec)

Set res = res + vec.

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

vec::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

res should be a vector, not a matrix.

vec should be a vector, not a matrix.

res and vec should have the same size

MuJoCo.mju_addToScl — Method

mju_addToScl(res, vec, scl)

Set res = res + vec*scl.

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

vec::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

scl::Float64

Additional Info

res should be a vector, not a matrix.

vec should be a vector, not a matrix.

res and vec should have the same size

MuJoCo.mju_band2Dense — Method

mju_band2Dense(res, mat, ntotal, nband, ndense, flg_sym)

Convert banded matrix to dense matrix, fill upper triangle if flg_sym>0.

Arguments

res::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

mat::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

ntotal::Int32

nband::Int32

ndense::Int32

flg_sym::UInt8

Additional Info

mat should be a vector, not a matrix.

res should have ntotal rows

res should have ntotal columns

MuJoCo.mju_bandMulMatVec — Method

mju_bandMulMatVec(res, mat, vec, ntotal, nband, ndense, nVec, flg_sym)

Multiply band-diagonal matrix with nvec vectors, include upper triangle if flg_sym>0.

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

mat::Matrix{Float64} -> A matrix variable size. Check additional info for sizes. Constant.

vec::Matrix{Float64} -> A matrix variable size. Check additional info for sizes. Constant.

ntotal::Int32

nband::Int32

ndense::Int32

nVec::Int32

flg_sym::UInt8

Additional Info

res should be a vector, not a matrix.

res should have ntotal rows

res should have nVec columns

vec should have ntotal rows

vec should have nVec columns

MuJoCo.mju_boxQP — Method

mju_boxQP(res, R, index, H, g, lower, upper)

minimize 0.5x'Hx + x'g s.t. lower <= x <= upper, return rank or -1 if failed inputs: n - problem dimension H - SPD matrix nn g - bias vector n lower - lower bounds n upper - upper bounds n res - solution warmstart n return value: nfree <= n - rank of unconstrained subspace, -1 if failure outputs (required): res - solution n R - subspace Cholesky factor nfreenfree allocated: n(n+7) outputs (optional): index - set of free dimensions nfree allocated: n notes: the initial value of res is used to warmstart the solver R must have allocatd size n(n+7), but only nfree*nfree values are used in output index (if given) must have allocated size n, but only nfree values are used in output only the lower triangles of H and R and are read from and written to, respectively the convenience function mju_boxQPmalloc allocates the required data structures

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

R::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

index::Vector{Int32} -> An optional vector of variable size. Check additional info for sizes.

H::Matrix{Float64} -> A matrix variable size. Check additional info for sizes. Constant.

g::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

lower::Vector{Float64} -> An optional vector of variable size. Check additional info for sizes. Constant.

upper::Vector{Float64} -> An optional vector of variable size. Check additional info for sizes. Constant.

Additional Info

res should be a vector, not a matrix.

index should be a vector, not a matrix.

g should be a vector, not a matrix.

lower should be a vector, not a matrix.

upper should be a vector, not a matrix.

size of R should be n*(n+7)

size of index should equal n

H should be of shape (n, n)

size of g should equal n

size of lower should equal n

size of upper should equal n

MuJoCo.mju_cholFactor — Method

mju_cholFactor(mat, mindiag)

Cholesky decomposition: mat = L*L'; return rank, decomposition performed in-place into mat.

Arguments

mat::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

mindiag::Float64

Additional Info

mat should be a square matrix

MuJoCo.mju_cholFactorBand — Method

mju_cholFactorBand(mat, ntotal, nband, ndense, diagadd, diagmul)

Band-dense Cholesky decomposition. Returns minimum value in the factorized diagonal, or 0 if rank-deficient. mat has (ntotal-ndense) x nband + ndense x ntotal elements. The first (ntotal-ndense) x nband store the band part, left of diagonal, inclusive. The second ndense x ntotal store the band part as entire dense rows. Add diagadd+diagmul*mat_ii to diagonal before factorization.

Arguments

mat::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

ntotal::Int32

nband::Int32

ndense::Int32

diagadd::Float64

diagmul::Float64

Additional Info

mat should be a vector, not a matrix.

MuJoCo.mju_cholSolve — Method

mju_cholSolve(res, mat, vec)

Solve (mat*mat') * res = vec, where mat is a Cholesky factor.

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

mat::Matrix{Float64} -> A matrix variable size. Check additional info for sizes. Constant.

vec::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

res should be a vector, not a matrix.

vec should be a vector, not a matrix.

mat should be a square matrix

MuJoCo.mju_cholSolveBand — Method

mju_cholSolveBand(res, mat, vec, ntotal, nband, ndense)

Solve (matmat')res = vec where mat is a band-dense Cholesky factor.

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

mat::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

vec::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

ntotal::Int32

nband::Int32

ndense::Int32

Additional Info

res should be a vector, not a matrix.

mat should be a vector, not a matrix.

vec should be a vector, not a matrix.

size of res should equal ntotal

size of vec should equal ntotal

MuJoCo.mju_cholUpdate — Method

mju_cholUpdate(mat, x, flg_plus)

Cholesky rank-one update: LL' +/- xx'; return rank.

Arguments

mat::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

x::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

flg_plus::Int32

Additional Info

x should be a vector, not a matrix.

mat should be a square matrix

MuJoCo.mju_copy — Method

mju_copy(res, data)

Set res = vec.

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

data::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

res should be a vector, not a matrix.

data should be a vector, not a matrix.

res and data should have the same size

MuJoCo.mju_d2n — Method

mju_d2n(res, vec)

Convert from double to mjtNum.

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

vec::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

res should be a vector, not a matrix.

vec should be a vector, not a matrix.

res and vec should have the same size

MuJoCo.mju_decodePyramid — Method

mju_decodePyramid(force, pyramid, mu)

Convert pyramid representation to contact force.

Arguments

force::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

pyramid::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

mu::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

force should be a vector, not a matrix.

pyramid should be a vector, not a matrix.

mu should be a vector, not a matrix.

MuJoCo.mju_dense2Band — Method

mju_dense2Band(res, mat, ntotal, nband, ndense)

Convert dense matrix to banded matrix.

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

mat::Matrix{Float64} -> A matrix variable size. Check additional info for sizes. Constant.

ntotal::Int32

nband::Int32

ndense::Int32

Additional Info

res should be a vector, not a matrix.

mat should have ntotal rows

mat should have ntotal columns

MuJoCo.mju_dot — Method

mju_dot(vec1, vec2)

Return dot-product of vec1 and vec2.

Arguments

vec1::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

vec2::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

vec1 should be a vector, not a matrix.

vec2 should be a vector, not a matrix.

vec1 and vec2 should have the same size

MuJoCo.mju_encodePyramid — Method

mju_encodePyramid(pyramid, force, mu)

Convert contact force to pyramid representation.

Arguments

pyramid::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

force::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

mu::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

pyramid should be a vector, not a matrix.

force should be a vector, not a matrix.

mu should be a vector, not a matrix.

MuJoCo.mju_eye — Method

mju_eye(mat)

Set mat to the identity matrix.

Arguments

mat::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

Additional Info

mat should be square

MuJoCo.mju_f2n — Method

mju_f2n(res, vec)

Convert from float to mjtNum.

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

vec::Vector{Float32} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

res should be a vector, not a matrix.

vec should be a vector, not a matrix.

res and vec should have the same size

MuJoCo.mju_fill — Method

mju_fill(res, val)

Set res = val.

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

val::Float64

Additional Info

res should be a vector, not a matrix.

MuJoCo.mju_insertionSort — Method

mju_insertionSort(res)

Insertion sort, resulting list is in increasing order.

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

Additional Info

res should be a vector, not a matrix.

MuJoCo.mju_insertionSortInt — Method

mju_insertionSortInt(res)

Integer insertion sort, resulting list is in increasing order.

Arguments

res::Vector{Int32} -> A vector of variable size. Check additional info for sizes.

Additional Info

res should be a vector, not a matrix.

MuJoCo.mju_isZero — Method

mju_isZero(vec)

Return 1 if all elements are 0.

Arguments

vec::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

Additional Info

vec should be a vector, not a matrix.

MuJoCo.mju_mulMatMat — Method

mju_mulMatMat(res, mat1, mat2)

Multiply matrices: res = mat1 * mat2.

Arguments

res::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

mat1::Matrix{Float64} -> A matrix variable size. Check additional info for sizes. Constant.

mat2::Matrix{Float64} -> A matrix variable size. Check additional info for sizes. Constant.

Additional Info

#rows in res should equal #rows in mat1

#columns in mat1 should equal #rows in mat2

MuJoCo.mju_mulMatMatT — Method

mju_mulMatMatT(res, mat1, mat2)

Multiply matrices, second argument transposed: res = mat1 * mat2'.

Arguments

res::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

mat1::Matrix{Float64} -> A matrix variable size. Check additional info for sizes. Constant.

mat2::Matrix{Float64} -> A matrix variable size. Check additional info for sizes. Constant.

Additional Info

#rows in res should equal #rows in mat1

#columns in res should equal #rows in mat2

MuJoCo.mju_mulMatTMat — Method

mju_mulMatTMat(res, mat1, mat2)

Multiply matrices, first argument transposed: res = mat1' * mat2.

Arguments

res::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

mat1::Matrix{Float64} -> A matrix variable size. Check additional info for sizes. Constant.

mat2::Matrix{Float64} -> A matrix variable size. Check additional info for sizes. Constant.

Additional Info

#rows in res should equal #columns in mat1

#rows in mat1 should equal #rows in mat2

MuJoCo.mju_mulMatTVec — Method

mju_mulMatTVec(res, mat, vec)

Multiply transposed matrix and vector: res = mat' * vec.

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

mat::Matrix{Float64} -> A matrix variable size. Check additional info for sizes. Constant.

vec::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

res should be a vector, not a matrix.

vec should be a vector, not a matrix.

MuJoCo.mju_mulMatVec — Method

mju_mulMatVec(res, mat, vec)

Multiply matrix and vector: res = mat * vec.

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

mat::Matrix{Float64} -> A matrix variable size. Check additional info for sizes. Constant.

vec::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

res should be a vector, not a matrix.

vec should be a vector, not a matrix.

MuJoCo.mju_mulVecMatVec — Method

mju_mulVecMatVec(vec1, mat, vec2)

Multiply square matrix with vectors on both sides: returns vec1' * mat * vec2.

Arguments

vec1::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

mat::Matrix{Float64} -> A matrix variable size. Check additional info for sizes. Constant.

vec2::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

vec1 should be a vector, not a matrix.

vec2 should be a vector, not a matrix.

MuJoCo.mju_n2d — Method

mju_n2d(res, vec)

Convert from mjtNum to double.

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

vec::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

res should be a vector, not a matrix.

vec should be a vector, not a matrix.

res and vec should have the same size

MuJoCo.mju_n2f — Method

mju_n2f(res, vec)

Convert from mjtNum to float.

Arguments

res::Vector{Float32} -> A vector of variable size. Check additional info for sizes.

vec::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

res should be a vector, not a matrix.

vec should be a vector, not a matrix.

res and vec should have the same size

MuJoCo.mju_norm — Method

mju_norm(vec)

Return vector length (without normalizing vector).

Arguments

vec::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

vec should be a vector, not a matrix.

MuJoCo.mju_normalize — Method

mju_normalize(vec)

Normalize vector, return length before normalization.

Arguments

vec::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

Additional Info

vec should be a vector, not a matrix.

MuJoCo.mju_printMat — Method

mju_printMat(mat)

Print matrix to screen.

Arguments

mat::Matrix{Float64} -> A matrix variable size. Check additional info for sizes. Constant.

MuJoCo.mju_scl — Method

mju_scl(res, vec, scl)

Set res = vec*scl.

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

vec::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

scl::Float64

Additional Info

res should be a vector, not a matrix.

vec should be a vector, not a matrix.

res and vec should have the same size

MuJoCo.mju_sqrMatTD — Method

mju_sqrMatTD(res, mat, diag)

Set res = mat' * diag * mat if diag is not NULL, and res = mat' * mat otherwise.

Arguments

res::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

mat::Matrix{Float64} -> A matrix variable size. Check additional info for sizes. Constant.

diag::Vector{Float64} -> An optional vector of variable size. Check additional info for sizes.

Additional Info

diag should be a vector, not a matrix.

#rows in res should equal #columns in mat

#rows in res should equal #columns in mat

MuJoCo.mju_sub — Method

mju_sub(res, vec1, vec2)

Set res = vec1 - vec2.

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

vec1::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

vec2::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

res should be a vector, not a matrix.

vec1 should be a vector, not a matrix.

vec2 should be a vector, not a matrix.

res and vec1 should have the same size

res and vec2 should have the same size

MuJoCo.mju_subFrom — Method

mju_subFrom(res, vec)

Set res = res - vec.

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

vec::Vector{Float64} -> A vector of variable size. Check additional info for sizes. Constant.

Additional Info

res should be a vector, not a matrix.

vec should be a vector, not a matrix.

res and vec should have the same size

MuJoCo.mju_sum — Method

mju_sum(vec)

Return sum(vec).

Arguments

vec::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

Additional Info

vec should be a vector, not a matrix.

MuJoCo.mju_symmetrize — Method

mju_symmetrize(res, mat)

Symmetrize square matrix res = (mat + mat')/2.

Arguments

res::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

mat::Matrix{Float64} -> A matrix variable size. Check additional info for sizes. Constant.

Additional Info

mat should be square

res and mat should have the same shape

MuJoCo.mju_transpose — Method

mju_transpose(res, mat)

Transpose matrix: res = mat'.

Arguments

res::Matrix{Float64} -> A matrix variable size. Check additional info for sizes.

mat::Matrix{Float64} -> A matrix variable size. Check additional info for sizes. Constant.

Additional Info

#columns in res should equal #rows in mat

#rows in res should equal #columns in mat

MuJoCo.mju_zero — Method

mju_zero(res)

Set res = 0.

Arguments

res::Vector{Float64} -> A vector of variable size. Check additional info for sizes.

Additional Info

res should be a vector, not a matrix.

MuJoCo.reset! — Method

reset!(m::Model, d::Data)

Resets the data values to their default states. You may equivalently use mj_resetData.

MuJoCo.resetkey! — Method

resetkey!(m::Model, d::Data, [keyframe = 1])

Resets the data struct to values in the supplied keyframe.

If no keyframe is specified, the first keyframe is used. The keyframe is a 1-based index into the list supplied by the model's specification.

MuJoCo.resetkey! — Method

Resets the data struct to values in the first key frame.

MuJoCo.sample_model_and_data — Method

sample_model_and_data()

A utility module to create and initialise example Model and Data objects, reflecting the underlying mjModel and mjData structs to provide REPL code completition to aid development.

Returns a (model, data) tuple.

MuJoCo.set_physics_state! — Method

set_physics_state!(m::Model, d::Data, x::AbstractVector)

Set the state vector of a MuJoCo model.

The state vector is [joint positions, joint velocities, actuator states] with dimension (nq, nv, na). Call forward! after setting the state to compute all derived quantities in the Data object according to the new state.

See also mj_setState and get_physics_state.

MuJoCo.step! — Method

step!(model::Model, data::Data)

Runs the simulation forward one time step, modifying the underlying data object.

MuJoCo.timestep — Method

timestep(model::MODEL_TYPES)

Extract the solver time-step for the given model.

Visualiser

MuJoCo.Visualiser.visualise!

MuJoCo.Visualiser.visualise! — Function

visualise!(m::Model, d::Data; controller=nothing)

Starts an interactive visualization of a MuJoCo model specified by an instance of Model and Data.

The visualizer has three "modes" that allow you to visualize passive dynamics, run a controller interactively, or play back recorded trajectories. The passive dynamics mode is always available, while the controller and trajectory modes are specified by the keyword arguments below.

Press F1 for help after running the visualiser to print the available mouse/button options in the terminal. Switch between modes with CTRL+RightArrow and CTRL+LeftArrow (or CMD for Mac). The different visualiser modes are ordered as follows:

Controller mode (if controller keyword is provided)

Trajectory mode (if trajectories keyword is provided)

Passive mode (always available)

Keywords

controller: a callback function with the signature controller(m, d), called at each timestep, that applies a control input to the system (or does any other operation you like).

trajectories: a single trajectory or Vector of trajectories, where each trajectory is an AbstractMatrix of states with size (nx, T) where nx = model.nq + model.nv + model.na and T is the length of the trajectory. Note that each trajectory can have a different length.

Examples

using MuJoCo
install_visualiser() # Run this to install dependencies only once
init_visualiser()    # Load required dependencies into session

# Load a model
model, data = MuJoCo.sample_model_and_data()

# Simulate and record a trajectory
T = 200
nx = model.nq + model.nv + model.na
states = zeros(nx, T)
for t in 1:T
    states[:,t] = get_physics_state(model, data)
    step!(model, data)
end

# Define a controller
function ctrl!(m,d)
    d.ctrl .= 2*rand(m.nu) .- 1
end

# Run the visualiser
reset!(model, data)
visualise!(model, data, controller=ctrl!, trajectories = states)

Named Access

MuJoCo.Wrappers.NamedAccess.actuator

MuJoCo.Wrappers.NamedAccess.body

MuJoCo.Wrappers.NamedAccess.cam

MuJoCo.Wrappers.NamedAccess.eq

MuJoCo.Wrappers.NamedAccess.exclude

MuJoCo.Wrappers.NamedAccess.geom

MuJoCo.Wrappers.NamedAccess.hfield

MuJoCo.Wrappers.NamedAccess.jnt

MuJoCo.Wrappers.NamedAccess.key

MuJoCo.Wrappers.NamedAccess.light

MuJoCo.Wrappers.NamedAccess.mat

MuJoCo.Wrappers.NamedAccess.mesh

MuJoCo.Wrappers.NamedAccess.numeric

MuJoCo.Wrappers.NamedAccess.pair

MuJoCo.Wrappers.NamedAccess.sensor

MuJoCo.Wrappers.NamedAccess.site

MuJoCo.Wrappers.NamedAccess.skin

MuJoCo.Wrappers.NamedAccess.tendon

MuJoCo.Wrappers.NamedAccess.tex

MuJoCo.Wrappers.NamedAccess.tuple

MuJoCo.Wrappers.NamedAccess.actuator — Function

actuator([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Index refers to MuJoCo IDs, which start at 0. Properties available are: (id, name, ctrl, length, moment, velocity, force)

MuJoCo.Wrappers.NamedAccess.body — Function

body([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Index refers to MuJoCo IDs, which start at 0. Properties available are: (id, name, applied, xpos, xquat, xmat, xipos, ximat, com, cinert, crb, cvel, linvel, angmom, cacc, int, ext)

MuJoCo.Wrappers.NamedAccess.cam — Function

cam([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Index refers to MuJoCo IDs, which start at 0. Properties available are: (id, name, xpos, xmat)

MuJoCo.Wrappers.NamedAccess.eq — Function

eq([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Index refers to MuJoCo IDs, which start at 0. Properties available are: (id, name, type, obj1id, obj2id, active, solref, solimp, data)

MuJoCo.Wrappers.NamedAccess.exclude — Function

exclude([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Index refers to MuJoCo IDs, which start at 0. Properties available are: (id, name, signature)

MuJoCo.Wrappers.NamedAccess.geom — Function

geom([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Index refers to MuJoCo IDs, which start at 0. Properties available are: (id, name, xpos, xmat)

MuJoCo.Wrappers.NamedAccess.hfield — Function

hfield([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Index refers to MuJoCo IDs, which start at 0. Properties available are: (id, name, size, nrow, ncol, adr)

MuJoCo.Wrappers.NamedAccess.jnt — Function

jnt([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Index refers to MuJoCo IDs, which start at 0. Properties available are: (id, name, xanchor, xaxis)

MuJoCo.Wrappers.NamedAccess.key — Function

key([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Index refers to MuJoCo IDs, which start at 0. Properties available are: (id, name, time, qpos, qvel, act, mpos, mquat)

MuJoCo.Wrappers.NamedAccess.light — Function

light([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Index refers to MuJoCo IDs, which start at 0. Properties available are: (id, name, xpos, xdir)

MuJoCo.Wrappers.NamedAccess.mat — Function

mat([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Index refers to MuJoCo IDs, which start at 0. Properties available are: (id, name, texid, texuniform, texrepeat, emission, specular, shininess, reflectance, rgba)

MuJoCo.Wrappers.NamedAccess.mesh — Function

mesh([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Index refers to MuJoCo IDs, which start at 0. Properties available are: (id, name, vertadr, vertnum, texcoordadr, faceadr, facenum, graphadr)

MuJoCo.Wrappers.NamedAccess.numeric — Function

numeric([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Index refers to MuJoCo IDs, which start at 0. Properties available are: (id, name, adr, size)

MuJoCo.Wrappers.NamedAccess.pair — Function

pair([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Index refers to MuJoCo IDs, which start at 0. Properties available are: (id, name, dim, geom1, geom2, signature, solref, solimp, margin, gap, friction)

MuJoCo.Wrappers.NamedAccess.sensor — Function

sensor([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Index refers to MuJoCo IDs, which start at 0. Properties available are: (id, name)

MuJoCo.Wrappers.NamedAccess.site — Function

site([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Index refers to MuJoCo IDs, which start at 0. Properties available are: (id, name, xpos, xmat)

MuJoCo.Wrappers.NamedAccess.skin — Function

skin([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Index refers to MuJoCo IDs, which start at 0. Properties available are: (id, name, matid, rgba, inflate, vertadr, vertnum, texcoordadr, faceadr, facenum, boneadr, bonenum)

MuJoCo.Wrappers.NamedAccess.tendon — Function

tendon([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Index refers to MuJoCo IDs, which start at 0. Properties available are: (id, name, wrapadr, wrapnum, Jrownnz, Jrowadr, J_colind, length, J, velocity)

MuJoCo.Wrappers.NamedAccess.tex — Function

tex([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Index refers to MuJoCo IDs, which start at 0. Properties available are: (id, name, type, height, width, adr)

MuJoCo.Wrappers.NamedAccess.tuple — Function

tuple([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Index refers to MuJoCo IDs, which start at 0. Properties available are: (id, name, adr, size)