MuJoCo API

MuJoCo

MuJoCo.example_model_files

MuJoCo.example_model_files_directory

MuJoCo.forward!

MuJoCo.get_physics_state

MuJoCo.humanoid_model_file

MuJoCo.init_data

MuJoCo.init_visualiser

MuJoCo.install_visualiser

MuJoCo.load_model

MuJoCo.mj_array

MuJoCo.mj_zeros

MuJoCo.reset!

MuJoCo.resetkey!

MuJoCo.resetkey!

MuJoCo.sample_model_and_data

MuJoCo.set_physics_state!

MuJoCo.step!

MuJoCo.timestep

MuJoCo.example_model_files — Method

Returns the list of absolute paths to available example model files provided by MuJoCo.

MuJoCo.example_model_files_directory — Method

Returns the absolute path of the root directory containing all the example model files.

MuJoCo.forward! — Method

forward!(model::MODEL_TYPES, data::DATA_TYPES)

Same as step! but without integrating in time.

MuJoCo.get_physics_state — Method

get_physics_state(m::Model, d::Data)

Extract the state vector of a MuJoCo model.

The state vector is [joint positions, joint velocities, actuator states] with dimension (nq, nv, na).

See also mj_getState and set_physics_state!.

MuJoCo.humanoid_model_file — Method

Returns the absolute path to the humanoid model provided by MuJoCo.

MuJoCo.init_data — Method

init_data(model::Model)

Creates an instance of the data required for the step! simulation.

Returns a Data object, wrapping the underlying mjData object.

MuJoCo.init_visualiser — Method

init_visualiser()

Loads the packages necessary for the running the visualiser.

Add the following packages to your project to be able to use the visualisation features, or run "install_visualiser()".

Packages:

BangBang

FFMPEG

GLFW

Observables

PrettyTables

Printf

StaticArrays

MuJoCo.install_visualiser — Method

install_visualiser()

Installs the necessary packages for the running the visualiser into the current running environment.

Packages:

BangBang

FFMPEG

GLFW

Observables

PrettyTables

Printf

StaticArrays

MuJoCo.load_model — Method

load_model(path)

Determines the type of file by the extension and loads the model into memory.

To use this model in a simulator, you will also need the corresponding data, obtained using init_data.

Expected files types: 'xml', 'mjcf', or 'mjb' (or uppercase variants).

Examples

model = load_model(MuJoCo.humanoid_model_file())
data = init_data(model)

MuJoCo.mj_array — Method

mj_array([element_type=mjtNum], dims...)

Allocates an array compatible with the underlying MuJoCo C API.

The C API treats arrays as row-major, while by default arrays in Julia are column-major. This function will create an array which is accessed in a row-major way, but can be treated by a normal array in your Julia code.

Arguments

element_type: Defaults to mjtNum (typically Float64).

dims: The dimensionality of output array. Can either be a tuple of integers or a series of integers.

MuJoCo.mj_zeros — Method

mj_zeros([element_type=mjtNum], dims...)

Allocates an array full of zeros, compatible with the underlying MuJoCo C API.

The C API treats arrays as row-major, while by default arrays in Julia are column-major. This function will create an array which is accessed in a row-major way, but can be treated by a normal array in your Julia code.

Arguments

element_type: The element type of the array. Defaults to mjtNum (typically Float64).

dims: The dimensionality of output array. Can either be a tuple of integers or a series of integers.

MuJoCo.reset! — Method

reset!(m::Model, d::Data)

Resets the data values to their default states. You may equivalently use mj_resetData.

MuJoCo.resetkey! — Method

resetkey!(m::Model, d::Data, [keyframe = 1])

Resets the data struct to values in the supplied keyframe.

If no keyframe is specified, the first keyframe is used. The keyframe is a 1-based index into the list supplied by the model's specification.

MuJoCo.resetkey! — Method

Resets the data struct to values in the first key frame.

MuJoCo.sample_model_and_data — Method

sample_model_and_data()

A utility module to create and initialise example Model and Data objects, reflecting the underlying mjModel and mjData structs to provide REPL code completition to aid development.

Returns a (model, data) tuple.

MuJoCo.set_physics_state! — Method

set_physics_state!(m::Model, d::Data, x::AbstractVector)

Set the state vector of a MuJoCo model.

The state vector is [joint positions, joint velocities, actuator states] with dimension (nq, nv, na). Call forward! after setting the state to compute all derived quantities in the Data object according to the new state.

See also mj_setState and get_physics_state.

MuJoCo.step! — Method

step!(model::Model, data::Data)

Runs the simulation forward one time step, modifying the underlying data object.

MuJoCo.timestep — Method

timestep(model::MODEL_TYPES)

Extract the solver time-step for the given model.

Visualiser

MuJoCo.Visualiser.visualise!

MuJoCo.Visualiser.visualise! — Function

visualise!(m::Model, d::Data; controller=nothing)

Starts an interactive visualization of a MuJoCo model specified by an instance of Model and Data.

The visualizer has three "modes" that allow you to visualize passive dynamics, run a controller interactively, or play back recorded trajectories. The passive dynamics mode is always available, while the controller and trajectory modes are specified by the keyword arguments below.

Press F1 for help after running the visualiser to print the available mouse/button options in the terminal.

Keywords

controller: a callback function with the signature controller(m, d), called at each timestep, that applies a control input to the system (or does any other operation you like).

trajectories: a single trajectory or Vector of trajectories, where each trajectory is an AbstractMatrix of states with size (nx, T) where nx = model.nq + model.nv + model.na and T is the length of the trajectory. Note that each trajectory can have a different length.

Examples

using MuJoCo
install_visualiser() # Run this to install dependencies only once
init_visualiser()    # Load required dependencies into session

# Load a model
model, data = MuJoCo.sample_model_and_data()

# Simulate and record a trajectory
T = 200
nx = model.nq + model.nv + model.na
states = zeros(nx, T)
for t in 1:T
    states[:,t] = get_physics_state(model, data)
    step!(model, data)
end

# Define a controller
function ctrl!(m,d)
    d.ctrl .= 2*rand(m.nu) .- 1
end

# Run the visualiser
reset!(model, data)
visualise!(model, data, controller=ctrl!, trajectories = states)

See also get_physics_state and reset!.

Named Access

MuJoCo.Wrappers.NamedAccess.actuator

MuJoCo.Wrappers.NamedAccess.body

MuJoCo.Wrappers.NamedAccess.cam

MuJoCo.Wrappers.NamedAccess.eq

MuJoCo.Wrappers.NamedAccess.exclude

MuJoCo.Wrappers.NamedAccess.geom

MuJoCo.Wrappers.NamedAccess.hfield

MuJoCo.Wrappers.NamedAccess.jnt

MuJoCo.Wrappers.NamedAccess.key

MuJoCo.Wrappers.NamedAccess.light

MuJoCo.Wrappers.NamedAccess.mat

MuJoCo.Wrappers.NamedAccess.mesh

MuJoCo.Wrappers.NamedAccess.numeric

MuJoCo.Wrappers.NamedAccess.pair

MuJoCo.Wrappers.NamedAccess.sensor

MuJoCo.Wrappers.NamedAccess.site

MuJoCo.Wrappers.NamedAccess.skin

MuJoCo.Wrappers.NamedAccess.tendon

MuJoCo.Wrappers.NamedAccess.tex

MuJoCo.Wrappers.NamedAccess.tuple

MuJoCo.Wrappers.NamedAccess.actuator — Function

actuator([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Properties available are: (id, name, ctrl, length, moment, velocity, force)

MuJoCo.Wrappers.NamedAccess.body — Function

body([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Properties available are: (id, name, applied, xpos, xquat, xmat, xipos, ximat, com, cinert, crb, cvel, linvel, angmom, cacc, int, ext)

MuJoCo.Wrappers.NamedAccess.cam — Function

cam([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Properties available are: (id, name, xpos, xmat)

MuJoCo.Wrappers.NamedAccess.eq — Function

eq([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Properties available are: (id, name, type, obj1id, obj2id, active, solref, solimp, data)

MuJoCo.Wrappers.NamedAccess.exclude — Function

exclude([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Properties available are: (id, name, signature)

MuJoCo.Wrappers.NamedAccess.geom — Function

geom([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Properties available are: (id, name, xpos, xmat)

MuJoCo.Wrappers.NamedAccess.hfield — Function

hfield([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Properties available are: (id, name, size, nrow, ncol, adr)

MuJoCo.Wrappers.NamedAccess.jnt — Function

jnt([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Properties available are: (id, name, xanchor, xaxis)

MuJoCo.Wrappers.NamedAccess.key — Function

key([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Properties available are: (id, name, time, qpos, qvel, act, mpos, mquat)

MuJoCo.Wrappers.NamedAccess.light — Function

light([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Properties available are: (id, name, xpos, xdir)

MuJoCo.Wrappers.NamedAccess.mat — Function

mat([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Properties available are: (id, name, texid, texuniform, texrepeat, emission, specular, shininess, reflectance, rgba)

MuJoCo.Wrappers.NamedAccess.mesh — Function

mesh([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Properties available are: (id, name, vertadr, vertnum, texcoordadr, faceadr, facenum, graphadr)

MuJoCo.Wrappers.NamedAccess.numeric — Function

numeric([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Properties available are: (id, name, adr, size)

MuJoCo.Wrappers.NamedAccess.pair — Function

pair([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Properties available are: (id, name, dim, geom1, geom2, signature, solref, solimp, margin, gap, friction)

MuJoCo.Wrappers.NamedAccess.sensor — Function

sensor([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Properties available are: (id, name)

MuJoCo.Wrappers.NamedAccess.site — Function

site([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Properties available are: (id, name, xpos, xmat)

MuJoCo.Wrappers.NamedAccess.skin — Function

skin([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Properties available are: (id, name, matid, rgba, inflate, vertadr, vertnum, texcoordadr, faceadr, facenum, boneadr, bonenum)

MuJoCo.Wrappers.NamedAccess.tendon — Function

tendon([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Properties available are: (id, name, wrapadr, wrapnum, Jrownnz, Jrowadr, J_colind, length, J, velocity)

MuJoCo.Wrappers.NamedAccess.tex — Function

tex([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Properties available are: (id, name, type, height, width, adr)

MuJoCo.Wrappers.NamedAccess.tuple — Function

tuple([model, data], [name, index])

Creates an object with access to views of the supplied model or data object, based either on an index or a name. Properties available are: (id, name, adr, size)